+++ List_more.v
@@ -923,8 +923,11 @@ Qed.
 
 (** ** [fold_right] *)
 
-Lemma fold_id A l : fold_right (@cons A) nil l = l.
+Lemma fold_right_id A l : fold_right (@cons A) nil l = l.
 Proof. induction l; [ reflexivity | cbn; f_equal; assumption ]. Qed.
+
+#[deprecated(since="ollibs 2.1.1", use=fold_right_id)]
+Definition fold_id := fold_right_id.
 
 Lemma fold_right_app_assoc2 A B f (g : B -> A) h (e : A) l1 l2 :
     (forall x y z, h (g x) (f y z) = f (h (g x) y) z) ->